// src/types.ts
export type Language = 'ar' | 'en' | 'ku' | 'kmr';

export type SimulationCluster = 
  | 'all'
  | 'cluster_a'
  | 'cluster_b'
  | 'cluster_c'
  | 'cluster_d'
  | 'cluster_e'
  | 'cluster_f'
  | 'atomic_quantum'
  | 'electricity_magnetism'
  | 'gravity_space'
  | 'energy_transformations'
  | 'advanced_waves'
  | 'mechanics_fluids';

export type ExperimentType = 
  // Cluster A: Atomic & Quantum Physics (1-7)
  | 'build_atom'
  | 'build_nucleus'
  | 'rutherford'
  | 'rutherford_scattering'
  | 'models_hydrogen'
  | 'models_h_atom'
  | 'blackbody'
  | 'blackbody_spectrum'
  | 'molecules_light'
  | 'molecules_and_light'
  | 'color_vision'
  // Cluster B: Electricity, Circuits & Magnetism (8-14)
  | 'capacitor'
  | 'capacitor_lab'
  | 'charges_fields'
  | 'charges_and_fields'
  | 'wire_resistance'
  | 'resistance_in_wire'
  | 'circuits'
  | 'circuit_construction_kit'
  | 'generator'
  | 'faradays_law'
  | 'magnet_compass'
  | 'magnetic_field'
  | 'magnets_electromagnets'
  // Cluster C: Gravity & Astrophysics (15-17)
  | 'gravity_orbits'
  | 'gravity_and_orbits'
  | 'kepler_laws'
  | 'keplers_laws'
  | 'projectile_motion'
  | 'gravity_force_lab'
  | 'solar_system'
  // Cluster D: Energy & Momentum (18-20)
  | 'collision_lab'
  | 'energy_skate_park'
  | 'pendulum_lab'
  | 'energy_forms'
  // Cluster E: Advanced Waves & Optics (21-25)
  | 'fourier_waves'
  | 'fourier_making_waves'
  | 'sound_waves'
  | 'normal_modes'
  | 'wave_on_string'
  | 'wave_on_a_string'
  | 'geometric_optics'
  | 'bending_light'
  // Cluster F: Mechanics, Fluids & Thermodynamics (26-30)
  | 'hookes_law_masses'
  | 'fluid_pressure_flow'
  | 'buoyancy_archimedes'
  | 'states_of_matter'
  | 'gas_diffusion'
  | 'diffusion'
  | 'balancing_act'
  | 'buoyancy'
  | 'forces_motion'
  | 'friction'
  | 'gas_properties'
  // Legacy / Additional
  | 'pendulum' 
  | 'projectile' 
  | 'optics' 
  | 'freefall' 
  | 'waves'
  | 'spring'
  | 'collision'
  | 'thermodynamics'
  | 'arc_length'
  | 'rotational_dynamics'
  | 'center_of_mass'
  | 'pendulum_energy'
  | 'acoustic_resonance'
  | 'sound_speed'
  | 'atomic_spectra'
  | 'metric_prefixes'
  | 'stress_strain'
  | 'bernoulli'
  | 'angled_mirrors'
  | 'curved_mirrors'
  | 'thin_lenses'
  | 'polarization'
  | 'light_scattering'
  | 'work_heat'
  | 'prescription_glasses'
  | 'periscope'
  | 'static_balloons'
  | 'sled_friction'
  | 'heat_conduction'
  | 'seesaw_torque'
  | 'electromagnetic_induction'
  | 'viscosity_stokes'
  | 'ramp_machine'
  // 5 New Experiments (IDs 66 to 70)
  | 'doppler_effect'
  | 'electrical_transformer'
  | 'photoelectric_effect'
  | 'radioactive_decay'
  | 'calorimetry_equilibrium';

export interface MeasurementRecord {
  id: string;
  experiment: ExperimentType;
  timestamp: string;
  parameters: Record<string, number | string>;
  measuredValue: number;
  theoreticalValue: number;
  unit: string;
  variableName: string;
  percentError: number;
  notes?: string;
  equation?: string;
}